+++ environments/gloo-mesh-core/additional-cluster-1/istio/helm/test.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 
-$SCRIPT_DIR/tools/wait-for-rollout.sh deployment istio-ingressgateway-1-23 istio-gateways 10 ${cluster_context}
+$SCRIPT_DIR/tools/wait-for-rollout.sh deployment istio-ingressgateway-1-24 istio-gateways 10 ${cluster_context}