+++ environments/gloo-mesh-core/singlecluster/istio/lifecyclemanager/test.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-$SCRIPT_DIR/tools/wait-for-rollout.sh deployment istio-operator-1-23 gm-iop-1-23 10 ${cluster_context}
-$SCRIPT_DIR/tools/wait-for-rollout.sh deployment istiod-1-23 istio-system 10 ${cluster_context}
-$SCRIPT_DIR/tools/wait-for-rollout.sh deployment istio-ingressgateway-1-23 istio-gateways 10 ${cluster_context}
-#./tools/wait-for-rollout.sh deployment istio-eastwestgateway-1-23 istio-eastwest 10 ${cluster_context}
+$SCRIPT_DIR/tools/wait-for-rollout.sh deployment istio-operator-1-24 gm-iop-1-24 10 ${cluster_context}
+$SCRIPT_DIR/tools/wait-for-rollout.sh deployment istiod-1-24 istio-system 10 ${cluster_context}
+$SCRIPT_DIR/tools/wait-for-rollout.sh deployment istio-ingressgateway-1-24 istio-gateways 10 ${cluster_context}
+#./tools/wait-for-rollout.sh deployment istio-eastwestgateway-1-24 istio-eastwest 10 ${cluster_context}