+++ environments/gloo-platform/gwapi-mgmt-gm-workers/cluster1/istio/lifecyclemanager/test.sh
@@ -3,7 +3,7 @@
 if [ "${environment_overlay}" == "ilcm" ] ; then 
      echo ""
   else 
-     $SCRIPT_DIR/tools/wait-for-rollout.sh deployment istio-ingressgateway-1-23 istio-gateways 10 ${cluster_context}
+     $SCRIPT_DIR/tools/wait-for-rollout.sh deployment istio-ingressgateway-1-24 istio-gateways 10 ${cluster_context}
      #$SCRIPT_DIR/tools/wait-for-rollout.sh deployment istio-eastwestgateway istio-gateways 10 ${cluster_context}
   fi